split header task into two tasks one for source files the other for dist - at build time only run header-dist - at publish time only commit version.js from the src - with a new year simply run header-src and it will not touch the dist
